Relocate executor code to executor package.
