+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/cloud/AppEngineDeploymentRuntime.java
@@ -16,6 +16,8 @@
 
 package com.google.cloud.tools.intellij.appengine.cloud;
 
+import com.google.cloud.tools.intellij.appengine.cloud.executor.AppEngineExecutor;
+import com.google.cloud.tools.intellij.appengine.cloud.executor.AppEngineStopTask;
 import com.google.cloud.tools.intellij.util.GctBundle;
 
 import com.intellij.icons.AllIcons.General;
@@ -35,7 +37,7 @@ import javax.swing.SwingUtilities;
 /**
  * Takes care of undeploy and stop of App Engine applications.
  */
-class AppEngineDeploymentRuntime extends DeploymentRuntime {
+public class AppEngineDeploymentRuntime extends DeploymentRuntime {
 
   private static final String STOP_CONFIRMATION_URI_OPEN_TAG =
       "<a href='https://cloud.google.com/appengine/docs/java/console/#versions'>";
